pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "elyth"
version = "0.0.0.2"
authors = [
  { name="Karl Benjamin R. Bughaw / ElythHQ", email="benjo@pro.space" }, 
]
description = "Dead-simple Discord bot creation for Python, inspired by BotForge." 
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" } 
keywords = ["discord", "bot", "discord.py", "simple", "beginner", "bdfd", "aoi.js"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License", # Adjust if needed
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
]
dependencies = [
    "discord.py>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/ElythHQ"
Repository = "https://github.com/ElythHQ/Elyth" 
"Bug Tracker" = "https://github.com/ElythHQ/Elyth/issues" 
[tool.setuptools.packages.find]
where = ["."]
include = ["elyth*"]
exclude = ["examples*", "tests*"]